tests: add withr to Suggests so that defer can be call in tests

+++ tests/testthat/setup.R
@@ -1,9 +1,11 @@
 old_cli_width <- options(cli.width = 200)
 
-withr::defer(
-  {
-    options(old_cli_width)
-    if (file.exists("Rplots.pdf")) file.remove("Rplots.pdf")
-  },
-  teardown_env()
-)
+if (require("withr")) {
+  withr::defer(
+    {
+      options(old_cli_width)
+      if (file.exists("Rplots.pdf")) file.remove("Rplots.pdf")
+    },
+    teardown_env()
+  )
+}
